Added basic Fastly integration
no issue
TODO: sort out bugs with replacing certificates and strip Fastly
library out of codebase. We should be using a shared Fastly library
between the cert-manager (which will handle out-of-band deletion of
unused certificates) and the greenlock store (which will handle the
rest of the flows relating to certificates).